third_party/blink/web_tests/svg/custom/multiple-title-elements-expected.txt

Expected , got : PASS
Expected First title, got First title: PASS
Expected First title, got First title: PASS
Expected Second title, got Second title: PASS
